requirements.txt
Django==4.1.7
djangorestframework==3.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
django-cors-headers==3.14.0
ipython==8.11.0
